Splitting held items to their own file
