+++ src/modifier/modifier-type.ts
@@ -20,14 +20,7 @@ import { getPokemonNameWithAffix } from "#app/messages";
 import {
   AddPokeballModifier,
   AddVoucherModifier,
-  AttackTypeBoosterModifier,
-  BaseStatModifier,
-  BerryModifier,
   BoostBugSpawnModifier,
-  BypassSpeedChanceModifier,
-  ContactHeldItemTransferChanceModifier,
-  CritBoosterModifier,
-  DamageMoneyRewardModifier,
   DoubleBattleChanceBoosterModifier,
   EnemyAttackStatusEffectChanceModifier,
   EnemyDamageBoosterModifier,
@@ -37,19 +30,15 @@ import {
   EnemyStatusEffectHealChanceModifier,
   EnemyTurnHealModifier,
   EvolutionItemModifier,
-  EvolutionStatBoosterModifier,
-  EvoTrackerModifier,
   ExpBalanceModifier,
   ExpBoosterModifier,
   ExpShareModifier,
   ExtraModifierModifier,
-  FlinchChanceModifier,
   FusePokemonModifier,
   GigantamaxAccessModifier,
   HealingBoosterModifier,
   HealShopCostModifier,
   HiddenAbilityRateBoosterModifier,
-  HitHealModifier,
   IvScannerModifier,
   LevelIncrementBoosterModifier,
   LockModifierTiersModifier,
@@ -60,44 +49,57 @@ import {
   MoneyRewardModifier,
   MultipleParticipantExpBonusModifier,
   PokemonAllMovePpRestoreModifier,
+  PokemonHpRestoreModifier,
+  PokemonLevelIncrementModifier,
+  PokemonNatureChangeModifier,
+  PokemonPpRestoreModifier,
+  PokemonPpUpModifier,
+  PokemonStatusHealModifier,
+  RememberMoveModifier,
+  ShinyRateBoosterModifier,
+  TempCritBoosterModifier,
+  TempStatStageBoosterModifier,
+  TerastallizeAccessModifier,
+  TerrastalizeModifier,
+  TmModifier,
+  type Modifier,
+  TempExtraModifierModifier,
+  CriticalCatchChanceBoosterModifier,
+} from "#app/modifier/modifier";
+import {
+  AttackTypeBoosterModifier,
+  BaseStatModifier,
+  BerryModifier,
+  BypassSpeedChanceModifier,
+  ContactHeldItemTransferChanceModifier,
+  CritBoosterModifier,
+  DamageMoneyRewardModifier,
+  EvolutionStatBoosterModifier,
+  EvoTrackerModifier,
+  FlinchChanceModifier,
+  HitHealModifier,
   PokemonBaseStatFlatModifier,
   PokemonBaseStatTotalModifier,
   PokemonExpBoosterModifier,
   PokemonFormChangeItemModifier,
   PokemonFriendshipBoosterModifier,
   PokemonHeldItemModifier,
-  PokemonHpRestoreModifier,
   PokemonIncrementingStatModifier,
   PokemonInstantReviveModifier,
-  PokemonLevelIncrementModifier,
   PokemonMoveAccuracyBoosterModifier,
   PokemonMultiHitModifier,
-  PokemonNatureChangeModifier,
   PokemonNatureWeightModifier,
-  PokemonPpRestoreModifier,
-  PokemonPpUpModifier,
-  PokemonStatusHealModifier,
   PreserveBerryModifier,
-  RememberMoveModifier,
   ResetNegativeStatStageModifier,
-  ShinyRateBoosterModifier,
   SpeciesCritBoosterModifier,
   SpeciesStatBoosterModifier,
   SurviveDamageModifier,
   SwitchEffectTransferModifier,
-  TempCritBoosterModifier,
-  TempStatStageBoosterModifier,
-  TerastallizeAccessModifier,
-  TerrastalizeModifier,
-  TmModifier,
   TurnHealModifier,
   TurnHeldItemTransferModifier,
   TurnStatusEffectModifier,
-  type Modifier,
-  TempExtraModifierModifier,
-  CriticalCatchChanceBoosterModifier,
   FieldEffectModifier,
-} from "#app/modifier/modifier";
+} from "#app/modifier/held-item-modifier";
 import { ModifierTier } from "#app/modifier/modifier-tier";
 import { getVoucherTypeIcon, getVoucherTypeName, VoucherType } from "#app/system/voucher";
 import type { PokemonMoveSelectFilter, PokemonSelectFilter } from "#app/ui/party-ui-handler";
