+++ src/ui/run-info-ui-handler.ts
@@ -27,6 +27,7 @@ import { SettingKeyboard } from "#app/system/settings/settings-keyboard";
 import { getBiomeName } from "#app/data/balance/biomes";
 import type { MysteryEncounterType } from "#enums/mystery-encounter-type";
 import { globalScene } from "#app/global-scene";
+import { modifierSortFunc, PokemonHeldItemModifier } from "#app/modifier/held-item-modifier";
 
 /**
  * RunInfoUiMode indicates possible overlays of RunInfoUiHandler.
@@ -650,7 +651,7 @@ export default class RunInfoUiHandler extends UiHandler {
       modifierIconsContainer.setScale(0.45);
       for (const m of this.runInfo.modifiers) {
         const modifier = m.toModifier(this.modifiersModule[m.className]);
-        if (modifier instanceof Modifier.PokemonHeldItemModifier) {
+        if (modifier instanceof PokemonHeldItemModifier) {
           continue;
         }
         const icon = modifier?.getIcon(false);
@@ -881,17 +882,17 @@ export default class RunInfoUiHandler extends UiHandler {
       const heldItemsScale =
         this.runInfo.gameMode === GameModes.SPLICED_ENDLESS || this.runInfo.gameMode === GameModes.ENDLESS ? 0.25 : 0.5;
       const heldItemsContainer = globalScene.add.container(-82, 2);
-      const heldItemsList: Modifier.PokemonHeldItemModifier[] = [];
+      const heldItemsList: PokemonHeldItemModifier[] = [];
       if (this.runInfo.modifiers.length) {
         for (const m of this.runInfo.modifiers) {
           const modifier = m.toModifier(this.modifiersModule[m.className]);
-          if (modifier instanceof Modifier.PokemonHeldItemModifier && modifier.pokemonId === pokemon.id) {
+          if (modifier instanceof PokemonHeldItemModifier && modifier.pokemonId === pokemon.id) {
             modifier.stackCount = m["stackCount"];
             heldItemsList.push(modifier);
           }
         }
         if (heldItemsList.length > 0) {
-          (heldItemsList as Modifier.PokemonHeldItemModifier[]).sort(Modifier.modifierSortFunc);
+          (heldItemsList as PokemonHeldItemModifier[]).sort(modifierSortFunc);
           let row = 0;
           for (const [index, item] of heldItemsList.entries()) {
             if (index > 36) {
