+++ src/ui/target-select-ui-handler.ts
@@ -6,7 +6,7 @@ import { getMoveTargets } from "../data/moves/move";
 import { Button } from "#enums/buttons";
 import type { Moves } from "#enums/moves";
 import type Pokemon from "#app/field/pokemon";
-import type { ModifierBar } from "#app/modifier/modifier";
+import type { ModifierBar } from "#app/modifier/modifier-bar";
 import { SubstituteTag } from "#app/data/battler-tags";
 import { globalScene } from "#app/global-scene";
